pyproject.toml
[project]
name = "notar-ia-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docxtpl>=0.20.2",
    "fastapi>=0.129.0",
    "google-api-python-client>=2.190.0",
    "google-auth>=2.48.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "httpx>=0.28.1",
    "openai>=2.21.0",
    "pdfplumber>=0.11.9",
    "pillow>=12.1.1",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.7.0",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.22",
    "rank-bm25>=0.2.2",
    "reportlab>=4.4.10",
    "tenacity>=9.1.4",
    "uvicorn[standard]>=0.40.0",
]
